+++ lib/kuby/tasks.rb
@@ -21,10 +21,17 @@ module Kuby
     end
 
     def build
+      build_args = {}
+
+      unless ENV.fetch('RAILS_MASTER_KEY', '').empty?
+        build_args['RAILS_MASTER_KEY'] = ENV['RAILS_MASTER_KEY']
+      end
+
       docker.cli.build(
         dockerfile: docker.to_dockerfile,
         image_url:  docker.metadata.image_url,
-        tags:       docker.metadata.tags
+        tags:       docker.metadata.tags,
+        build_args: build_args
       )
     end